Update darkmodelib to 0.21.2
(cherry picked from commit 32b285e3060b3fb296b2d1fe4ccb8226ffc96e04)

+++ Src/Common/MDITabBar.cpp
@@ -75,7 +75,7 @@ BOOL CMyTabCtrl::Create(CMDIFrameWnd* pMainFrame, CWnd* pParent)
 		DarkMode::setTabCtrlUpDownSubclass(hSelf);
 
 	if (HWND hTip = m_tooltips.GetSafeHwnd())
-		DarkMode::setDarkTooltips(hTip, DarkMode::ToolTipsType::tooltip);
+		DarkMode::setDarkTooltips(hTip, static_cast<int>(DarkMode::ToolTipsType::tooltip));
 
 	return TRUE;
 }
